Moved Search Component to Components Folder
