+++ client/src/App.js
@@ -1,12 +1,9 @@
 import React from 'react';
 
 import './App.css';
-import Search from './Search'
+import Search from './Components/Search'
 
 class App extends React.Component {
-  
-  
-  
   render() {
   return (
     <div className="App">
