--- client/src/Search.module.css
@@ -1,30 +1,0 @@
-
-.search #submitForm{
-    position: relative;
-    top:250px;
-    display:flex;
-    justify-content: center;
-   
-  }
-.search input{
-    height: 40px;
-    size:40px;
-  }
-.search  button{
-    margin:5px;
-    left:10px;
-    position:relative;
-    background:35px;
-    color:white;
-    border-radius: 10px;
-    height:40px;
-  }
-  .search #title{
-    text-align: center;
-    position: relative;
-    top:100px;
-    color:white;
-  }
-  
-  
-
